Modified read speed
Reduced delay times from 250ms and 20ms to 1ms.  This was done in accordance with the datasheet.  E.G. no where does it say you need to wait 250ms or 20ms.  It does say to set the line high for a minimum of 1ms (the second delay).  This has been tested and is working in my of my projects