Use micros() instead of cycle counting.
In the ESP32 Arduino library, noInterrupts() is a nop, so interrupts
still occur while receiving a value. On my board there is at least one
interrupt every time we try to read a value, which messes up the
reading. Instead of counting cycles, use micros() to keep track of time.
This increases the success rate of receiving data to 90% or so.

A better fix would be to implement noInterrupts() for the ESP32, but
that looks a lot harder. The new code path is probably not suitable for
slower micros, but hopefully those use the __AVR code path anyway.